Reusinage du helper de rdv
